pal: update number of channels  in resourcemanager for proxy
When we get device switch back to back without setparameters the value
of channels is 0 because of which the channel number in resource manager
is not getting updated. Hence, updating the number of channels and reset
is done on device close when no device is connected.

Change-Id: I54031d57587df4b8b1deed3a9e9e09f3880c5833
(cherry picked from commit bbd5bae9668e2a48ca703a0ae8f3f59d5a450d09)