PAL: Remove size check, as it was specific to DOA usecase
Size check was specific to DOA usecase, but size of DOA payload
was harcoded, hence we don't need a size check for this, and this
size check was restricting ASR get param, as its payload size was
greater than DOA payload size.

Removed the size check, to enable ASR get param.

Change-Id: I77d7d77a005ed72a4a84e9323291a5cdc7e83ca6