ar-pal: remove querying current tx profile for SensorPCMData
For SensorPCMData stream, when trying to get the current
capture profile, no need to compare with existing profile
attributes, as this can be handled in update sound trigger
capture profile.
It also fixes a deadlock issue, in updateSoundTriggerProfile
with rm lock held, it can call GetTXMacroCaptureProfile for
SensorPCMData, which needs to hold rm lock again, leading
to deadlock.

Change-Id: I041b3f43b68fe831ab5ee359fd047e6b3fbb2f2c