pal: switch device to a2dp for orphan stream in a2dpreconfig
Sometimes switching device to a2dp fails because of BT controller error.
but APM thought stream is running on a2dp, so a2dp reconfig comes
instead of device switch cmd after that, but no active stream is running
on a2dp device so reconfig fails and muting happens.

Check if there is active orphan stream which has empty device list, but
pal device is a2dp, that means it ever failed to switch to a2dp,
this kind of stream is needed to be switched to a2dp in a2dpreconfig.

Change-Id: I53feee40772b05ef86da1a379adbf8776432df76