pal: Set exit_buffering_ to False after Device switch
In DualVA usecase while second stage detection is being
processed and a device switch is requested.
After Device switch is handled, any subsequent attempt
for second stage detection fails as the flag exit_buffering_
remains true.

In the above change, additional check is handled in the
Connect Event of the Last stream to check for any Active
Stream and if found then set the flag to false.

Change-Id: I0eb1de069ab74285969003652a8ffb3aee8124fe